cross_agentic_hub/pyproject.toml
[project]
name = "cross-agentic-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.5",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
